fix incorrect variable replacement for messageSubject in featureSelector; add checkHandler and checkSubject to LiveMessagesUtil class and run argument not null checks in LiveMessagesUtil;
